Module calcu_rs::abstract_alg

Macros

tuple_op_list_impls 🔒

Structs

Group

A group is an Algebra < G; ⋆, ^, e > satisfying the following axioms:
G1: ⋆ is associative
G2: e is a neutral element (e.g Identity)
G3: ∀ a ∈ G has an Inverse element a ⋆ â = â ⋆ a = e

Monoid

A monoid is an Algebra < M; ⋆, e> where ⋆ is an associative and e is the neutral element

Traits

Algebra

An Algebra is a pair < S; Ω >, where S is a set and Ω = (ω1, …, ωn) is a list of Operations on S

Associative

a ⋆ (b ⋆ c) = (a ⋆ b) ⋆ c

Binary

binary Operation

Commutative

a ⋆ b = b ⋆ a

Distributive

a ⋆ (b + c) = (a ⋆ b) + (a ⋆ c)

Identity

A neutral element of < S; ⋆ >
e ∈ S: e ⋆ a = a ⋆ e = a ∀ a ∈ S

Inverse

An inverse element a in an Algebra < S; ⋆, e > is an element b ∈ S such that: b ⋆ a = a ⋆ b = e

LInverse

left Inverse, gets auto implemented when Inverse is implemented

Operation

An Operation on a set S is a function Sn → S, where n ≥ 0 is called the *arity

OperatorList

A tuple of Operation used for Algebra

RInverse

left Inverse, gets auto implemented when Inverse is implemented

Set

a set

Ternary

ternary Operation

Unary

unary Operation